+++ internal/css_printer/css_printer_test.go
@@ -48,9 +48,9 @@ func expectPrintedMinify(t *testing.T, contents string, expected string) {
 	})
 }
 
-func expectPrinteMangleSyntax(t *testing.T, contents string, expected string) {
+func expectPrintedMangle(t *testing.T, contents string, expected string) {
 	t.Helper()
-	expectPrintedCommon(t, contents+" [minified]", contents, expected, Options{
+	expectPrintedCommon(t, contents+" [mangled]", contents, expected, Options{
 		MangleSyntax: true,
 	})
 }
@@ -409,16 +409,20 @@ func TestASCII(t *testing.T) {
 
 func TestEmptyRule(t *testing.T) {
 	expectPrinted(t, "div {}", "div {\n}\n")
-	expectPrinted(t, ".a {}", ".a {\n}\n")
-	expectPrinted(t, ".a {}.b { color: red;}", ".a {\n}\n.b {\n  color: red;\n}\n")
-	expectPrinted(t, "@media (prefers-reduced-motion: no-preference) {\n}", "@media (prefers-reduced-motion: no-preference) {\n}\n")
-	expectPrinted(t, "@keyframes test {\n  from {}\n  to {\n  color: red;\n}\n}", "@keyframes test {\n  from {\n  }\n  to {\n    color: red;\n  }\n}\n")
+	expectPrinted(t, "$invalid {}", "$invalid {\n}\n")
+	expectPrinted(t, "@media screen {}", "@media screen {\n}\n")
+	expectPrinted(t, "@keyframes test { from {} to { color: red } }", "@keyframes test {\n  from {\n  }\n  to {\n    color: red;\n  }\n}\n")
+	expectPrinted(t, "@keyframes test { from { color: red } to {} }", "@keyframes test {\n  from {\n    color: red;\n  }\n  to {\n  }\n}\n")
 
-	expectPrintedMinify(t, "p:hover {}", "p:hover{}")
-	expectPrintedMinify(t, ".a {}\n.b {\n  color: red;\n}", ".a{}.b{color:red}")
-	expectPrintedMinify(t, "@keyframes mini {\n  from {}\n  to {\n  color: red;\n}\n}", "@keyframes mini{from{}to{color:red}}")
+	expectPrintedMinify(t, "div {}", "div{}")
+	expectPrintedMinify(t, "$invalid {}", "$invalid{}")
+	expectPrintedMinify(t, "@media screen {}", "@media screen{}")
+	expectPrintedMinify(t, "@keyframes test { from {} to { color: red } }", "@keyframes test{from{}to{color:red}}")
+	expectPrintedMinify(t, "@keyframes test { from { color: red } to {} }", "@keyframes test{from{color:red}to{}}")
 
-	expectPrinteMangleSyntax(t, "a:hover {}", "")
-	expectPrinteMangleSyntax(t, ".a {}\n.b {\n  color: red;\n}", ".b {\n  color: red;\n}\n")
-	expectPrinteMangleSyntax(t, "@keyframes test {\n  from {}\n  to {\n  color: red;\n}\n}", "@keyframes test {\n  to {\n    color: red;\n  }\n}\n")
+	expectPrintedMangle(t, "div {}", "")
+	expectPrintedMangle(t, "$invalid {}", "$invalid {\n}\n")
+	expectPrintedMangle(t, "@media screen {}", "")
+	expectPrintedMangle(t, "@keyframes test { from {} to { color: red } }", "@keyframes test {\n  to {\n    color: red;\n  }\n}\n")
+	expectPrintedMangle(t, "@keyframes test { from { color: red } to {} }", "@keyframes test {\n  from {\n    color: red;\n  }\n}\n")
 }